Move schedule generation api to app module
