+++ backend-application/src/main/java/eu/socialedge/hermes/backend/application/api/ScheduleGenerationApi.java
@@ -12,9 +12,9 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  */
-package eu.socialedge.hermes.backend.schedule.api;
+package eu.socialedge.hermes.backend.application.api;
 
-import eu.socialedge.hermes.backend.schedule.ResourceMappingsSupport;
+import eu.socialedge.hermes.backend.application.util.ResourceMappingsSupport;
 import eu.socialedge.hermes.backend.schedule.gen.BasicScheduleGenerator;
 import eu.socialedge.hermes.backend.schedule.repository.ScheduleRepository;
 import eu.socialedge.hermes.backend.transit.domain.Line;
@@ -23,7 +23,6 @@ import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.webmvc.RepositoryRestController;
 import org.springframework.http.ResponseEntity;
-import org.springframework.util.AntPathMatcher;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.util.UriComponentsBuilder;
